requirements.txt
psycopg2-binary
python-dotenv
pyyaml
kubernetes
docker